Make changes to allow for custom number of reflections to be included above 128

+++ src/MtzdumpHandling.jl
@@ -6,7 +6,7 @@ using PyCall
 ########################## Types Declarations ###########################
 immutable MtzdumpParams
     inputFilename::ASCIIString
-    nref::Int8
+    nref::Int64
 end
 
 #########################################################################
